feat: Make the image rootless and running headless
Currently the image crashes when first run because the taskd-init script has to be run.
This change introduces an entrypoint script that wraps taskd-init and runs it if some
init file is not found.

Moreover, taskd does not need to run as root, so a low-privileged user is used instead.
To do this, a multi-step build is created. As a side effect, the final image size is
reduced from 140MB to 27.8MB.